Move generalised FunctionsController to new lib
A little bit of jiggery-pokery is required here. First, the two parts of
the FunctionsController class have the `partial` modifier removed;
second, the public methods GetFunctionsOutput() and TeardownFunctions()
are reimplemented as delegating members on the SpecFlow-specific
version; finally, there's a little bit of type-system hackery in
DemoFunctionPerFeatureHooks to convince the compiler to import the
appropriate version of FunctionController. This last can be easily
resolved by renaming one or other of the classes as preferred.
